Added TaskServiceBase.aadl. Contains AGREE spec based on the wiki spec and refinement with Derek. The model differs slightly from the code.
Difference 1: made changes to some data event port types based on comments from Derek that will be realized in the code soon.

Difference 2: added additional ports not in the parent task code to meet the high-level task spec. These ports are actually implemented for certain children tasks only; however, in the spec the use of these ports are based on the high-level task state machine model. Adding these ports at this level allows for realizability checks on the complete task spec within this file.

This version represents a draft with comments that will need to be reviewed, resulting in further refinement and spec clean up.