fully support offline builds, and improve Meson files
The main work of this patch is to allow any dependency which
previously used a wrap file to also check for a suitable
system-provided dependency. In most cases, Meson will query
`pkg-config` for the dependency, but there are some exceptions for
dependencies with non-standard packaging like cppzmq and SQLiteCpp.

This patch also contains some elimination of redundancies between the
various `meson.build` files, particularly around how we specify
`cpp_args` for the sub-libraries.